requirements.txt
google_auth_oauthlib==0.4.2
dataclasses==0.8
google_api_python_client==1.12.8
protobuf==3.14.0
